shift transaction add to separate page
