+++ src/pages/planning/index.tsx
@@ -6,6 +6,8 @@ import { useSelector } from 'starfx/react';
 import IcicleChart from './icicleChart';
 import Transactions from './transactions';
 import Accounts from './accounts';
+import { Button } from '~/src/elements/Button';
+import { useNavigate } from 'react-router-dom';
 
 const Planning = (props) => {
   const accounts = useSelector(schema.accounts.selectTableAsList);
@@ -17,7 +19,7 @@ const Planning = (props) => {
           <div className="bg-white shadow rounded-lg divide-y divide-gray-200">
             <IcicleChart data={model.state.transactionsComputed} />
           </div>
-          <Divider text="Transactions" />
+          <Divider text="Transactions" navigateTo="/transactions/set" />
           <Transactions />
           <Divider text="Accounts" />
           <Accounts accounts={accounts} />
@@ -29,16 +31,25 @@ const Planning = (props) => {
 
 export default Planning;
 
-const Divider = ({ text }) => {
+const Divider = ({
+  text,
+  navigateTo
+}: {
+  text: string;
+  navigateTo: string;
+}) => {
+  const navigate = useNavigate();
   return (
     <div className="relative my-8">
       <div className="absolute inset-0 flex items-center" aria-hidden="true">
         <div className="w-full border-t border-gray-300" />
       </div>
-      <div className="relative flex justify-center">
+      <div className="relative flex justify-between">
+        <span />
         <span className="px-3 bg-gray-50 text-lg font-medium text-gray-900">
           {text}
         </span>
+        <Button onPress={() => navigate(navigateTo)}>Add</Button>
       </div>
     </div>
   );
