+++ src/pages/transactions/transactionInputAmountComputed.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Field, FieldArray } from 'formik';
-import { FieldGroup, Label } from '~/src/components/Form';
+import { FieldGroup, Label } from '~/src/components/Form.js';
 import { Button } from '~/src/elements/Button';
 import { Input } from '~/src/elements/Input';
 import { Select } from '~/src/elements/Select';
